refactor: Move contest table to features (#3169)
